feat(data-table): implement useColumnsFixedOffsets for managing fixed column offsets
